Rename folder of Cypress commands related to API endpoints
